Support for "pass-through" BSON serialization using Salat
This allows to store messages as BSON instead of a binary blob in MongoDB, which can be useful for several reasons, such as being able to perform queries on the journal in MongoDB directly.

It is possible to add custom BSON encoders / decoders if the default ones are not suffiscient.

The implementation does not make use of the default akka serialization mechanism as the aim is to store the message as BSON and the akka serialization mechanism has binary as a target format.